# Replace unbounded code conversion cache with limited LRU cache