This PR replaces the former unbound cache for converted codes in the LFVM interpreter with a fixed-sized LRU cache. Additionally, some unnecessary debug code was removed.

This fixes #314.

The PR also adds benchmarks for the cache to identify the performance impact:
```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/vm/lfvm
cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
                              │   old.txt   │               new.txt               │
                              │   sec/op    │   sec/op     vs base                │
ConversionCacheLookupSpeed-12   31.61n ± 4%   51.20n ± 4%  +61.97% (p=0.000 n=10)
ConversionCacheUpdateSpeed-12   573.1n ± 6%   427.7n ± 6%  -25.38% (p=0.000 n=10)
geomean                         134.6n        148.0n        +9.94%

                              │   old.txt    │               new.txt                │
                              │     B/op     │    B/op     vs base                  │
ConversionCacheLookupSpeed-12   0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
ConversionCacheUpdateSpeed-12   318.0 ± 8%     112.0 ± 0%  -64.78% (p=0.000 n=10)
geomean                                    ²               -40.65%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                              │   old.txt    │            new.txt             │
                              │  allocs/op   │ allocs/op   vs base            │
ConversionCacheLookupSpeed-12   0.000 ± 0%     0.000 ± 0%  ~ (p=1.000 n=10) ¹
ConversionCacheUpdateSpeed-12   0.000 ± 0%     1.000 ± 0%  ? (p=0.000 n=10)
geomean                                    ²               ?                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

By changing to the utilization of an LRU instead of the previous map, an ~62% increase in the lookup time is observed. This is likely due to the additional need to track the least-recently-used order. However, as shown in the following result, this does not have any significant impact on the interpreter performance when running codes.
```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/vm/test
cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
               │   old.txt   │              new.txt               │
               │   sec/op    │   sec/op     vs base               │
Inc/1/lfvm-12    3.373µ ± 5%   3.327µ ± 9%       ~ (p=0.393 n=10)
Inc/10/lfvm-12   3.263µ ± 7%   3.389µ ± 4%       ~ (p=0.123 n=10)
geomean          3.317µ        3.357µ       +1.21%
```
The `Inc` stress-tests the interpreter startup and shutdown time by running a simple code merely increasing a given value by a fixed offset and returning the result.

Remaining tasks:
- [x] check end-to-end impact using micro benchmarks
- [x] test potential impact on CT tests
- [x] do a test-run on historic data