@LuisPH3 2024-05-29
The PR replaces unbound cache with bound cache. This is used during the convert phase of the lfvm which runs once before code execution. Test with repeated initializations stresses this convert phase. Nevertheless, the instruction set in the program may not stress this cache entirely.